// apps/tourism/platform/src/types/rajatrain.ts
export interface definitions {
// ...
  rajaTariff:
    | "TARIFF_UNSPECIFIED"
    | "TARIFF_ADULT"
    | "TARIFF_CHILD"
    | "TARIFF_EMPTY"
    | "TARIFF_INFANT"
    | "TARIFF_ADULT_FOREIGN"
    | "TARIFF_CHILD_FOREIGN"
    | "TARIFF_INFANT_FOREIGN";
// ...
  rajaTrainTicket: {
    id?: string;
    ticketStatus?: definitions["rajaTicketStatus"];
    /** Format: uint64 */
    price?: string;
    passenger?: definitions["rajaPassenger"];
    /** Format: int64 */
    issueDate?: string;
    tariff?: definitions["rajaTariff"];
    seatNumber?: string;
    compartmentNumber?: string;
    wagonNumber?: string;
    option?: definitions["rajaOptionalService"][];
    freeOption?: definitions["rajaOptionalService"][];
    ticketNumber?: string;
    ticketSeries?: string;
    penaltyAmount?: string;
    penaltyPercentage?: string;
    refundAmount?: string;
    persianStatus?: string;
    /** Format: uint64 */
    totalPrice?: string;
    trainId?: string;
  };
// ...
}